\name{PictureSummary-class}
\docType{class}
\alias{PictureSummary-class}

\title{Class "PictureSummary" }
\description{Summary information about a picture
such as the number of paths and bounding box information.}
\section{Slots}{
	 \describe{
    \item{\code{numPaths}:}{Object of class \code{"numeric"} }
    \item{\code{xscale}:}{Object of class \code{"numeric"} 
range of x-values in picture.}
    \item{\code{yscale}:}{Object of class \code{"numeric"} 
range of y-values in picture.}
  }
}
\author{ Paul Murrell }
\seealso{
\code{\link{Picture-class}}
}
\keyword{classes}
